Revert "Somewhat added a feature"
