--- CodeFactory/Assets/Art/Ores/Materials/CoalOre.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9f94070e79dbb8247889789de2055800
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
