--- CodeFactory/Assets/Art/Ores/Materials/DiamondOre.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 921c36d05af29724dbeaf551b0492f97
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
