--- CodeFactory/Assets/Art/Ores/Materials/GoldOre.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 408447d57255d7345b1d7acd0fb2f8a3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
