--- CodeFactory/Assets/Art/Ores/Materials/IronOre.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e5d0a8f1b24d5854e87640aaea8dd664
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
