+++ CodeFactory/Assets/Prefabs/Cube.prefab
@@ -15,8 +15,6 @@ GameObject:
   - component: {fileID: -4661629614611016741}
   - component: {fileID: 8355207647066490351}
   - component: {fileID: -3562063368714491383}
-  - component: {fileID: 2333859263841877054}
-  - component: {fileID: -95819810606362527}
   m_Layer: 0
   m_Name: Cube
   m_TagString: ore
@@ -119,11 +117,11 @@ Rigidbody:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 9051482793227874658}
   serializedVersion: 2
-  m_Mass: 1e+9
+  m_Mass: 20
   m_Drag: 0
   m_AngularDrag: 0.05
   m_UseGravity: 1
-  m_IsKinematic: 0
+  m_IsKinematic: 1
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
@@ -140,51 +138,3 @@ SphereCollider:
   serializedVersion: 2
   m_Radius: 1.2425108
   m_Center: {x: -0.21606374, y: 0.36794424, z: 0}
---- !u!114 &2333859263841877054
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9051482793227874658}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b4a7156d9b4c86b449b4a18177f67075, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  script: {fileID: -95819810606362527}
-  Ore: {fileID: 9051482793227874658}
-  oreClosest: {fileID: 0}
-  Player: {fileID: 0}
-  stoneMaterial: {fileID: 2100000, guid: 2152566ceb5f2bd448a701c151e36fac, type: 2}
-  coalMaterial: {fileID: 2100000, guid: 9f94070e79dbb8247889789de2055800, type: 2}
-  ironMaterial: {fileID: 2100000, guid: e5d0a8f1b24d5854e87640aaea8dd664, type: 2}
-  goldMaterial: {fileID: 2100000, guid: 408447d57255d7345b1d7acd0fb2f8a3, type: 2}
-  diamondMaterial: {fileID: 2100000, guid: 921c36d05af29724dbeaf551b0492f97, type: 2}
-  ranNum: 0
---- !u!114 &-95819810606362527
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9051482793227874658}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 41a8aeb24357d07459737997c5ccf369, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ore: []
-  orePrefab: {fileID: 0}
-  oreExplode: {fileID: 0}
-  sparkPrefab: {fileID: 0}
-  oreClosest: {fileID: 0}
-  blowup: 0
-  ore_detection: 0
-  collectOre: 0
-  radius: 5
-  power: 175
-  Player: {fileID: 0}
-  firerate: 1
-  nextfire: 1
-  sparks: {fileID: 0}
