+++ CodeFactory/Assets/Scripts/OreMining.cs
@@ -1,4 +1,4 @@
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -7,7 +7,6 @@ public class OreMining : MonoBehaviour
     // These variables hold assets
     public GameObject[] ore;
     public GameObject orePrefab;
-
     public GameObject oreExplode;
     public GameObject sparkPrefab;
     public GameObject oreClosest;
@@ -35,12 +34,49 @@ public class OreMining : MonoBehaviour
     // Update is called once per frame
 
 
-    void FixedUpdate()
+     void FixedUpdate()
     {
+        // Destroys ore when orebot mines ore
+        // Debug.Log(blowup);
+        // Debug.Log(oreClosest);
         if (blowup == true && oreClosest != null)
         {
+            // Gets Player tag to add ores to inventory
+            Player = GameObject.FindGameObjectWithTag("Player");
+
+            // Gets the Player's inventory by using the Player tag and assigns it as a variable
+            var playerInv = Player.GetComponent<Inventory>();
+
+            // Ore Randomization Start
+            // Makes a number from 1 to 100
             if (oreClosest.GetComponent<OreHealth>().health <= 0)
             {
+                var ranNum = Random.Range(0, 100) + 1;
+
+                // Uses the randomly selected number to chose an ore and add it to the player's inventory
+                if (ranNum <= 50)
+                {
+                    playerInv.stoneCount += 1;
+                }
+                if (ranNum > 50 && ranNum <= 73)
+                {
+                    playerInv.coalCount += 1;
+                }
+                if (ranNum > 73 && ranNum <= 91)
+                {
+                    playerInv.ironCount += 1;
+                }
+                if (ranNum > 91 && ranNum < 100)
+                {
+                    playerInv.goldCount += 1;
+                }
+                if (ranNum == 100)
+                {
+                    playerInv.diamondCount += 1;
+                }
+                // Ore Randomization End
+
+
                 // Replaces the destroyed ore with an ore exploding animation. Starts Here.
                 var cloneOreExplode = Instantiate(oreExplode, oreClosest.gameObject.transform.position, Quaternion.identity); // Clones the oreExplode asset so I can delete the cloned asset instead of the original asset
                 Vector3 explosionPos = oreClosest.gameObject.transform.position; // Sets the coordinates of the explosion
@@ -74,5 +110,4 @@ public class OreMining : MonoBehaviour
     }
 
 }
-
