--- CodeFactory/Assets/Scripts/OreRandom.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4a7156d9b4c86b449b4a18177f67075
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
