+++ CodeFactory/Assets/Scripts/WalkBotScript.cs
@@ -219,6 +219,7 @@ public class WalkBotScript : MonoBehaviour
     }
     private void OnTriggerExit(Collider other)
     {
+
         this.GetComponent<OreMining>().blowup = false;
     }
 }
